Some problems I came across
Thanks for the great article! It's a really nice explanation of the Metropolis-Hastings algorithm and of how to do sampling in Computational Bayesian Statistics! I came across some small (but important) problems in the code. 
Most notably - when there is more than one data point, the likelihood function needs to be the product of the likelihoods of each individual data point - which I added above to the code for the functions. Perhaps you would want to mention this differently in the text.
Another problem is that you need to sample the unbounded sigma value from the previous *unbounded* sigma value, rather than the bounded one.
A third small correction - the pdf in the text above stands for density rather than distribution, as your title suggests, but the text itself used the word distribution, which is incorrect in this context.

I really enjoyed your article and hope my small proposed changes will help improve it a bit further :)

Omri